requirements.txt
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
joblib==1.5.1
narwhals==1.48.1
numpy==2.2.6
packaging==25.0
pandas==2.3.1
plotly==6.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
spotipy==2.25.1
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
